+++ src/main/java/com/example/active/business/controller/CategoryController.java
@@ -4,9 +4,11 @@ import com.example.active.business.domain.CategoryDTO;
 import com.example.active.business.service.CategoryService;
 import com.example.active.data.entity.Category;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -20,7 +22,8 @@ public class CategoryController {
 
     @Autowired
     private CategoryService categoryService;
-
+    @Autowired
+    private ApiKeyAuthenticator apiKeyAuthenticator;
     @RequestMapping(
             value = "/categories",
             method = RequestMethod.GET,
@@ -28,7 +31,7 @@ public class CategoryController {
     )
     @ResponseStatus(HttpStatus.OK)
     public List<CategoryDTO> getCategories(
-            @RequestParam(name = "key") String apiKey,
+            @RequestParam(name = "key", required = false) String apiKey,
             @RequestParam(name = "q", defaultValue = "") String query,
             HttpServletRequest request,
             HttpServletResponse response
@@ -44,7 +47,7 @@ public class CategoryController {
     @ResponseStatus(HttpStatus.OK)
     public Optional<CategoryDTO> getCategoryById(
             @PathVariable("id") Long id,
-            @RequestParam(name = "key") String apiKey,
+            @RequestParam(name = "key", required = false) String apiKey,
             @RequestParam(name = "q", defaultValue = "") String query,
             HttpServletRequest request,
             HttpServletResponse response
@@ -59,7 +62,7 @@ public class CategoryController {
     @ResponseStatus(HttpStatus.OK)
     public List<CategoryDTO> getCategoriesByFacility(
             @PathVariable("facility") String facility,
-            @RequestParam(name = "key") String apiKey,
+            @RequestParam(name = "key", required = false) String apiKey,
             @RequestParam(name = "q", defaultValue = "") String query,
             HttpServletRequest request,
             HttpServletResponse response
@@ -80,22 +83,31 @@ public class CategoryController {
             HttpServletResponse response,
             @RequestBody Category category
     ){
+        if(!apiKeyAuthenticator.isAuthenticated(apiKey)){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You cannot consume this service. Please check your api key");
+        }
         categoryService.save(category);
     }
 
     @RequestMapping(
             value = "/categories/{id}",
             method = RequestMethod.DELETE,
-            produces = {"application/json"},
-            consumes = {"application/json"}
+            consumes = {"*/*"}
     )
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void saveAvailableActivities(
+    public void deleteCategory(
             @RequestParam(name = "key") String apiKey,
             @PathVariable("id") Long id,
             HttpServletRequest request,
             HttpServletResponse response
     ){
-        categoryService.deleteCategory(id);
+        if(!apiKeyAuthenticator.isAuthenticated(apiKey)){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You cannot consume this service. Please check your api key");
+        }
+        try{
+            categoryService.deleteCategory(id);
+        } catch (EmptyResultDataAccessException e){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, String.format("No category found with id %d", id));
+        }
     }
 }
