+++ src/main/java/com/example/active/business/controller/TypeController.java
@@ -6,8 +6,10 @@ import com.example.active.business.service.FacilityService;
 import com.example.active.business.service.TypeService;
 import com.example.active.data.entity.Type;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -18,16 +20,16 @@ import java.util.List;
 public class TypeController {
     @Autowired
     private TypeService typeService;
-
     @Autowired
     private FacilityService facilityService;
-
+    @Autowired
+    private ApiKeyAuthenticator apiKeyAuthenticator;
     @RequestMapping(value = "",
             method = RequestMethod.GET,
             produces = {"application/json"})
     @ResponseStatus(HttpStatus.OK)
     public List<TypeDTO> getTypes(
-            @RequestParam(name = "key") String apiKey,
+            @RequestParam(name = "key", required = false) String apiKey,
             @RequestParam(name = "q", defaultValue = "") String query,
             @RequestParam(name = "sort") String sort,
             HttpServletRequest request,
@@ -42,7 +44,7 @@ public class TypeController {
     @ResponseStatus(HttpStatus.OK)
     public List<FacilityDTO> getFacilitiesByType(
             @PathVariable("type") String type,
-            @RequestParam(name = "key") String apiKey,
+            @RequestParam(name = "key", required = false) String apiKey,
             @RequestParam(name = "sort") String sort,
             @RequestParam(name = "q", defaultValue = "") String query,
             @RequestParam(name = "lng") Double lng,
@@ -66,12 +68,15 @@ public class TypeController {
             HttpServletResponse response,
             @RequestBody Type type
     ){
+        if(!apiKeyAuthenticator.isAuthenticated(apiKey)){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You cannot consume this service. Please check your api key");
+        }
         typeService.save(type);
     }
 
     @RequestMapping(value = "/{id}",
             method = RequestMethod.DELETE,
-            produces = {"application/json"})
+            consumes = {"*/*"})
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteType(
             @PathVariable("id") Long id,
@@ -79,6 +84,13 @@ public class TypeController {
             HttpServletRequest request,
             HttpServletResponse response
     ){
-        typeService.deleteType(id);
+        if(!apiKeyAuthenticator.isAuthenticated(apiKey)){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You cannot consume this service. Please check your api key");
+        }
+        try{
+            typeService.deleteType(id);
+        } catch (EmptyResultDataAccessException e){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, String.format("No type found with id %d", id));
+        }
     }
 }
